Restructure and refactor Card component
