+++ src/features/planner/components/index.tsx
@@ -2,18 +2,12 @@
 
 import { DayInfo, PlaceT } from "@/types";
 import { parseOrder } from "@/utils";
-import {
-  createDay,
-  updateDay,
-  updateDayOrder,
-  updatePlaceOrder,
-  updateStartTime,
-} from "@/utils/actions";
+import { updatePlaceOrder, updateStartTime } from "@/utils/actions/crud/update";
 import { add, format, parse } from "date-fns";
 import { Reorder } from "framer-motion";
+import { useRouter } from "next/navigation";
 import { useEffect, useState } from "react";
 import Card from "./Card";
-import { useRouter } from "next/navigation";
 import NavigateDays from "./NavigateDays";
 
 type Props = {
